requirements.txt
anthropic>=0.42.0
openai>=1.0.0
langchain>=0.3.0
chromadb>=0.5.0
faiss-cpu>=1.8.0
arxiv>=2.1.0
semanticscholar>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
